+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bump runtimeVersion to `2025.06.20`.
+ * Upgraded stable Flutter analysis SDK to `3.32.4`.
 
 ## `20250619t085100-all`
  * Bump runtimeVersion to `2025.06.03`.